Restructure the project to support external libraries distributed with the source. Add Howard Hinnant's Date library. Implement timestamps for logging. This closes #5.

+++ engine/engine.cpp
@@ -200,7 +200,7 @@ void Engine::RunGameLoop() {
   }
 }
 
-void Engine::Update(unsigned int delta_time_micros) {
+void Engine::Update(unsigned long long delta_time_micros) {
   timing->PassUpdateTimingInfo(delta_time_micros);
 
   // Update all engine subsystems in forward order.
